Big Bundle of Flats
Most of the Flats that I have (nearly all the ones I could find floating free in folders).

There are likely a few I missed to find on a second pass of my files + 1 or 2 factions I may have skipped. Many of my flats will need to be obtained through the xcf's (like for Shell enemies) so those will come later as well.